requirements.txt
albumentations==1.3.1
appdirs==1.4.4
argcomplete==3.1.2
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.8.0
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.21.1
icecream==2.1.3
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
lit==17.0.6
MarkupSafe==2.1.5
matplotlib==3.8.0
mpmath==1.3.0
munch==2.5.0
networkx==3.2.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.0
pipx==1.2.1
pretrainedmodels==0.7.4
protobuf==4.25.3
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
segmentation-models-pytorch==0.3.3
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torch_geometric==2.3.1
torchmetrics==1.0.3
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.28.1
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
userpath==1.9.1
wandb==0.15.12
